Update 10/12/2021 Like - Dislike & Security

+++ back/controllers/sauces.js
@@ -4,7 +4,7 @@ const fs = require("fs");
 
 //Function to display all sauces
 module.exports.getAllSauce = (req, res, next) => {
-  //we use "find" method from mangoose to find sauces in our DB
+  //we use "find" method from mongoose to find sauces in our DB
   Sauce.find()
     //Response is sended
     .then((sauces) => res.status(200).json(sauces))
@@ -14,7 +14,7 @@ module.exports.getAllSauce = (req, res, next) => {
 
 //Function do display one sauce
 module.exports.getOneSauce = (req, res, next) => {
-  //we use findOne method from mangoose to find only ONE sauce in our DB
+  //we use findOne method from mongoose to find only ONE sauce in our DB
   Sauce.findOne({ _id: req.params.id })
     //Response is sended
     .then((sauce) => res.status(200).json(sauce))
@@ -80,8 +80,9 @@ module.exports.deleteSauce = (req, res, next) => {
       
       //With her name, we can now remove the sauce
       fs.unlink(`images/${filename}`, () => {
-        //Then we delete the sauce with "deleteOne" property in MangoDB
-        Sauce.deleteOne({ _id: req.params.id })
+        //Then we delete the sauce with "deleteOne" property in MongoDB
+        //double check, sauce Id & User Id
+        Sauce.deleteOne({ _id: req.params.id }, { deleteSauce: req.body.userId})
           //Positive response ==> code 200
           .then(() => res.status(200).json({ message: "Holé la sauce a été supprimée !" }))
           //Negative response ==> code 400
@@ -90,4 +91,71 @@ module.exports.deleteSauce = (req, res, next) => {
     })
     //If don't, we catch an error message
     .catch((error) => res.status(500).json({ error }));
+};
+
+//Here we handle the like / dislike
+module.exports.likesDislikes = (req, res, next) => {
+  //In here, valor of "1" is he to indicated a "like"
+  if(req.body.like === 1) {
+      //Here we use "updateOne" to make our sauces updated
+      Sauce.updateOne(
+          //Update sauce is defined by her ID
+          {_id: req.params.id},
+          //"$ink" parameter of MongoDB to make 1 more like
+          //"$push" parameter of MongoDB to append our value in the array
+          {$inc: {likes: +1}, $push: { usersLiked: req.body.userId}})
+          //Positive response ==> Code 200
+          .then(() => res.status(200).json({message : "Holé j'aime cette sauce !"}))
+          //Negative response ==> Code 400 error message
+          .catch(error => res.status(400).json({ error }))
+  }
+  //Here, the negative value "-1" has the signification of one dislike
+  else if(req.body.like === -1) {
+      //"updateOne" is here to update our sauce's likes
+      Sauce.updateOne(
+          //the sauce is define by her "ID"
+          {_id: req.params.id},
+          //Here we use  $inc of MongoDB to increment our dislikes by 1
+          {$inc: {dislikes: +1}, $push: {usersDisliked: req.body.userId}})
+          //Then we send back positive response ==> code 200
+          .then((sauce) => res.status(200).json({message : "Holé Je n'aime plus cette sauce !"}))
+          //If don't we throw a negative response ==> code 400
+          .catch(error => res.status(400).json({ error }))
+
+  }
+  //If the value is equal to 0, this is the signification of cancelation of one like or dislike
+  else {
+      //So we find our sauce by her "ID"
+      Sauce.findOne({_id: req.params.id})
+      .then(sauce => {
+          //In the case user is in the array of values
+          if(sauce.usersLiked.includes(req.body.userId)) {
+              //Sauce update "updateOne"
+              Sauce.updateOne(
+                  //We first remove the user from the array of likes by using "$pull" of MongoDB
+                  {$pull: {usersLiked: req.body.userId,
+                  //Then we remothe the like
+                  $inc: {likes: -1}}})
+                  //Positive response ==> code 200
+                  .then(() => res.status(200).json({message: "Holé J'aime enlevé"}))
+                  //Negative response ==> code 400
+                  .catch(error => res.status(400).json({error})
+              )
+          }
+          //Else, if the user is the arry of dislikes
+          else if(sauce.usersDisliked.includes(req.body.userId)) {
+              //Sauce update "updateOne"
+              Sauce.updateOne(
+                  //We first remove the user from the array of dislikes
+                  {$pull: {usersDisliked: req.body.userId, 
+                  //then we remove the dislike
+                  $inc: {Dislikes: -1}}})
+                  //Positive reponse ==> code 200
+                  .then(() => res.status(200).json({message: "Holé Dislike enlevé !"}))
+                  //Negative resonse ==> code 400
+                  .catch(error => res.status(400).json({error})
+              )
+          }
+      })
+  }
 };
